+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to the [Semantic Versioning 2.0 Specification](https://semv
 ### ⚠️ Breaking
 
 - JSON, JSONC, and JSON5 are no longer supported as data file formats.
+- Supplying a custom template renderer is no longer supported.
 
 ## 0.2.0 - February 01, 2023